ai-service-api/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
numpy
pillow
scikit-learn
torch
torchvision
tensorflow
opencv-python
ultralytics
albumentations
python-dotenv
httpx
loguru
pytest
black